Split DAO tests into separate files
for readability.
