+++ scripts/test/db/startup.sh
@@ -45,7 +45,7 @@ fi
 # create ramdisk using diskutil
 echo "Creating ramdisk..."
 sudo mkdir ${RAMDISK_MOUNT_POINT}
-RAMDISK_DEVICE=$(sudo mount -t tmpfs -o size=800m tmpfs ${RAMDISK_MOUNT_POINT})
+RAMDISK_DEVICE=$(sudo mount -t tmpfs -o size=1g tmpfs ${RAMDISK_MOUNT_POINT})
 echo "$RAMDISK_DEVICE" | sudo tee ${RAMDISK_DEVICE_FILE}
 
 # give 'vagrant' user/group ownership of the RAM-disk folder.
