+++ tests/jest/db/CategoryDAO.spec.js
@@ -1,0 +1,20 @@
+import { StudyType } from '@/lib/Constants';
+import CategoryDAO from '@/lib/db/CategoryDAO';
+import Category from '@/lib/model/Category';
+
+test('CategoryDAO', async () => {
+  expect(CategoryDAO.isInited()).toBe(false);
+
+  let category = await CategoryDAO.byId(1);
+  expect(category.id).toBe(1);
+  expect(category.studyType).toBe(StudyType.ATR_VOLUME);
+  await expect(
+    Category.read.validateAsync(category),
+  ).resolves.toEqual(category);
+  expect(CategoryDAO.isInited()).toBe(true);
+
+  category = await CategoryDAO.byId(-1);
+  expect(category).toBeUndefined();
+
+  await expect(CategoryDAO.all()).resolves.toBeInstanceOf(Map);
+});
