+++ tests/jest/db/CollisionFactorDAO.spec.js
@@ -1,0 +1,14 @@
+import CollisionFactorDAO from '@/lib/db/CollisionFactorDAO';
+
+test('CollisionFactorDAO', async () => {
+  expect(CollisionFactorDAO.isInited()).toBe(false);
+
+  const collisionFactors = await CollisionFactorDAO.all();
+  expect(collisionFactors).toBeInstanceOf(Map);
+  expect(collisionFactors.get('acclass')).toBeInstanceOf(Map);
+  expect(collisionFactors.get('acclass').get(1)).toEqual({
+    code: 'FA',
+    description: 'Fatal',
+  });
+  expect(CollisionFactorDAO.isInited()).toBe(true);
+});
